bat: Split bat.go
The bat.go file was getting a little large.  It makes sense to split
it up into separate files, one per functional area, before any new
code gets added.
